backend/djoser-master/requirements.txt
djangorestframework
django
-e .
django-templated-mail==1.1.1
djangorestframework-simplejwt==4.3.0
djet==0.2.2
social-auth-app-django==3.1.0
coverage==4.5.3
coreapi

pytest==5.0.1
pytest-cov==2.7.1
pytest-django==3.5.1
pytest-pythonpath==0.7.3

